Added: Builder for file move

+++ tests/test_builders.py
@@ -4,6 +4,7 @@
 
 import os
 import shutil
+import time
 from pathlib import Path
 
 from ward import test, fixture, raises, xfail
@@ -29,7 +30,7 @@ def checkTmpFile():
 
 
 @fixture
-def setupTestCopy():
+def setupTestCopyMove():
     # Create some test files and directories
     os.makedirs("/tmp/remake", exist_ok=True)
     os.chdir("/tmp/remake")
@@ -38,12 +39,14 @@ def setupTestCopy():
     test_dir_1 = "test_dir_1"
     test_dir_2 = "test_dir_2"
     test_dir_3 = "test_dir_3"
+    test_dir_4 = "test_dir_4"
     test_file_3 = test_dir_1 + "/test_file_3.txt"
     link_to_file_1 = "link_to_file_1.lnk"
     link_to_file_3 = "link_to_file_3.lnk"
     os.makedirs(test_dir_1, exist_ok=True)
     os.makedirs(test_dir_2, exist_ok=True)
     os.makedirs(test_dir_3, exist_ok=True)
+    os.makedirs(test_dir_4, exist_ok=True)
     with open(test_file_1, "w") as f:
         f.write("This is a test file.")
     with open(test_file_2, "w") as f:
@@ -56,14 +59,24 @@ def setupTestCopy():
     yield
 
     # Clean up test files and directories
-    os.remove(test_file_1)
-    os.remove(test_file_2)
-    os.remove(test_file_3)
-    os.remove(link_to_file_1)
-    os.remove(link_to_file_3)
-    shutil.rmtree(test_dir_1)
-    shutil.rmtree(test_dir_2)
-    shutil.rmtree(test_dir_3)
+    if os.path.isfile(test_file_1):
+        os.remove(test_file_1)
+    if os.path.isfile(test_file_2):
+        os.remove(test_file_2)
+    if os.path.isfile(test_file_3):
+        os.remove(test_file_3)
+    if os.path.islink(link_to_file_1):
+        os.remove(link_to_file_1)
+    if os.path.islink(link_to_file_3):
+        os.remove(link_to_file_3)
+    if os.path.isdir(test_dir_1):
+        shutil.rmtree(test_dir_1)
+    if os.path.isdir(test_dir_2):
+        shutil.rmtree(test_dir_2)
+    if os.path.isdir(test_dir_3):
+        shutil.rmtree(test_dir_3)
+    if os.path.isdir(test_dir_4):
+        shutil.rmtree(test_dir_4)
 
 
 @test("Builders can handle python functions")
@@ -122,9 +135,9 @@ def test_04_buildersKwargs():
     getCurrentContext().clearRules()
 
 
-@test("Basic copy operations.")
-def test_05_copyFileOperations(_=setupTestCopy):
-    """Basic copy operations."""
+@test("Basic file copy operations.")
+def test_05_copyFileOperations(_=setupTestCopyMove):
+    """Basic file copy operations."""
     def _doCopy(src, dst):
         getCurrentContext().clearRules()
         Rule(deps=src, targets=dst, builder=cp).apply()
@@ -134,12 +147,18 @@ def test_05_copyFileOperations(_=setupTestCopy):
     test_file_2 = Path("test_file_2.txt")
     test_dir_1 = Path("test_dir_1")
     test_dir_2 = Path("test_dir_2")
-    test_dir_3 = Path("test_dir_3")
-    test_file_3 = test_dir_1 / "test_file_3.txt"
-    link_to_file_1 = Path("link_to_file_1.lnk")
-    link_to_file_3 = Path("link_to_file_3.lnk")
 
     # Single file
+    source = test_file_1
+    destination = test_dir_1
+    _doCopy(source, destination)
+    assert (destination / source).is_file() is True
+    with open(destination / source, "r") as f:
+        content = f.read()
+    assert content == "This is a test file."
+    os.remove(destination / source)
+
+    # Single file with rename
     source = test_file_1
     destination = test_dir_1 / "copied_file.txt"
     _doCopy(source, destination)
@@ -147,6 +166,20 @@ def test_05_copyFileOperations(_=setupTestCopy):
     with open(destination, "r") as f:
         content = f.read()
     assert content == "This is a test file."
+    os.remove(destination)
+
+    # Single file with replace
+    source = test_file_1
+    destination = test_dir_1 / "renamed_file.txt"
+    _doCopy(source, destination)
+    source = test_file_2
+    time.sleep(0.1)
+    source.touch()
+    _doCopy(source, destination)
+    assert destination.exists() is True
+    with open(destination, "r") as f:
+        content = f.read()
+    assert content == "Another test file."
     os.remove(destination)
 
     # Multiple files
@@ -165,16 +198,48 @@ def test_05_copyFileOperations(_=setupTestCopy):
     with raises(FileNotFoundError):
         _doCopy(source, destination)
 
-    # Non-existent destination directory
+    # Non-existent destination directory single file
     source = test_file_1
     destination = "nonexistent_dir/copied_file.txt"
-    with raises(OSError):
+    with raises(FileNotFoundError):
         _doCopy(source, destination)
+
+    # Non-existent destination directory
+    source = [test_file_1, test_file_2]
+    destination = "nonexistent_dir"
+    with raises(FileNotFoundError):
+        _doCopy(source, destination)
+
+
+@test("Basic directory copy operations.")
+def test_06_copyDirectoryOperations(_=setupTestCopyMove):
+    """Basic directory copy operations."""
+    def _doCopy(src, dst):
+        getCurrentContext().clearRules()
+        Rule(deps=src, targets=dst, builder=cp).apply()
+        getCurrentContext().clearRules()
+
+    test_dir_1 = Path("test_dir_1")
+    test_dir_2 = Path("test_dir_2")
+    test_dir_3 = Path("test_dir_3")
+    test_dir_4 = Path("test_dir_4")
+    test_file_3 = test_dir_1 / "test_file_3.txt"
 
     # Single directory
     source = test_dir_1
-    destination_dir = test_dir_3 / "copied_dir"
+    destination_dir = test_dir_3
     destination_file = destination_dir / test_file_3
+    _doCopy(source, destination_dir)
+    assert destination_file.exists() is True
+    assert destination_file.is_file() is True
+    assert (destination_dir / source).exists() is True
+    assert (destination_dir / source).is_dir() is True
+    shutil.rmtree(destination_dir / source)
+
+    # Single directory with rename
+    source = test_dir_1
+    destination_dir = test_dir_3 / "renamed_dir"
+    destination_file = destination_dir / test_file_3.name
     _doCopy(source, destination_dir)
     assert destination_file.exists() is True
     assert destination_file.is_file() is True
@@ -184,9 +249,8 @@ def test_05_copyFileOperations(_=setupTestCopy):
 
     # Multiple directories
     sources = [test_dir_1, test_dir_2]
-    destination_dir = test_dir_3 / "copied_dirs"
+    destination_dir = test_dir_3
     destination_file = destination_dir / test_file_3
-    # destination_file = test_file_3.replace("test_dir_1", destination_dir)
     _doCopy(sources, destination_dir)
     assert destination_file.exists() is True
     assert destination_file.is_file() is True
@@ -202,9 +266,23 @@ def test_05_copyFileOperations(_=setupTestCopy):
     with raises(FileNotFoundError):
         _doCopy(source, destination)
 
-    # Non-existent destination directory for directory
+    # Non-existent destination directory
     # OK: Will work since shutil.copytree will create intermediate directories.
-    # Thus same as aabove.
+    # Thus same as above.
+
+
+@test("Basic link copy operations.")
+def test_07_copyLinkOperations(_=setupTestCopyMove):
+    """Basic link copy operations."""
+    def _doCopy(src, dst):
+        getCurrentContext().clearRules()
+        Rule(deps=src, targets=dst, builder=cp).apply()
+        getCurrentContext().clearRules()
+
+    test_dir_1 = Path("test_dir_1")
+    test_dir_2 = Path("test_dir_2")
+    link_to_file_1 = Path("link_to_file_1.lnk")
+    link_to_file_3 = Path("link_to_file_3.lnk")
 
     # IMPORTANT: Links are automatically resolved during rule process!
     # Single link
@@ -241,15 +319,108 @@ def test_05_copyFileOperations(_=setupTestCopy):
         _doCopy(source, destination)
 
 
-@test("Basic move operations.")
-@xfail
-def test_06_moveFileOperations():
-    """Basic move operations."""
-    raise NotImplementedError
-    # === Move ===
-    # Files (single, multiple, non-existant source, non-existant dest)
-    # Directorie (single, multiple, non-existant source, non-existant dest)
-    # Links (single, multiple, non-existant source, non-existant dest)
+@test("Basic file move operations.")
+def test_08_moveFileOperations(_=setupTestCopyMove):
+    """Basic file move operations."""
+    def _doMove(src, dst):
+        getCurrentContext().clearRules()
+        Rule(deps=src, targets=dst, builder=mv).apply()
+        getCurrentContext().clearRules()
+
+    test_file_1 = Path("test_file_1.txt")
+    test_file_2 = Path("test_file_2.txt")
+    test_dir_1 = Path("test_dir_1")
+    test_dir_2 = Path("test_dir_2")
+    test_file_3 = test_dir_1 / "test_file_3.txt"
+
+    # Single file
+    source = test_file_1
+    destination = test_dir_1 / "moved_file.txt"
+    _doMove(source, destination)
+    assert source.exists() is False
+    assert destination.exists() is True
+    with open(destination, "r") as f:
+        content = f.read()
+    assert content == "This is a test file."
+    os.remove(destination)
+
+    # Multiple files
+    sources = [test_file_2, test_file_3]
+    destination_dir = test_dir_2
+    _doMove(sources, destination_dir)
+    for source in sources:
+        filename = os.path.basename(source)
+        destination = destination_dir / filename
+        assert source.exists() is False
+        assert destination.exists() is True
+        os.remove(destination)
+
+    # Non-existant source file
+    source = "nonexistent_file.txt"
+    destination = test_dir_1 / "moved_file.txt"
+    with raises(FileNotFoundError):
+        _doMove(source, destination)
+
+    # Non-existent destination directory
+    source = test_file_1
+    destination = "nonexistent_dir/moved_file.txt"
+    with raises(OSError):
+        _doMove(source, destination)
+
+
+@test("Basic directory move operations.")
+def test_09_moveDirectoryOperations(_=setupTestCopyMove):
+    """Basic directory move operations."""
+    def _doMove(src, dst):
+        getCurrentContext().clearRules()
+        Rule(deps=src, targets=dst, builder=mv).apply()
+        getCurrentContext().clearRules()
+
+    test_dir_1 = Path("test_dir_1")
+    test_dir_2 = Path("test_dir_2")
+    test_dir_3 = Path("test_dir_3")
+    test_dir_4 = Path("test_dir_4")
+    test_file_1 = Path("test_file_1.txt")
+    test_file_2 = Path("test_file_2.txt")
+
+    # Single directory
+    source = test_dir_1
+    destination_dir = test_dir_3
+    _doMove(source, destination_dir)
+    assert source.exists() is False
+    assert destination_dir.exists() is True
+    assert destination_dir.is_dir() is True
+
+    # Single directory with renaming
+    source = test_dir_2
+    destination_dir = test_dir_3 / "moved_dir"
+    _doMove(source, destination_dir)
+    assert source.exists() is False
+    assert destination_dir.exists() is True
+    assert destination_dir.is_dir() is True
+
+    # Multiple directories
+    sources = [test_dir_3, test_file_1]
+    destination_dir = test_dir_4
+    # breakpoint()
+    _doMove(sources, destination_dir)
+    assert all([_.exists() is False for _ in sources])
+    assert (destination_dir / sources[0]).exists() is True
+    assert (destination_dir / sources[0]).is_dir() is True
+    assert (destination_dir / sources[1]).exists() is True
+    assert (destination_dir / sources[1]).is_file() is True
+
+    # Non-existent source directory
+    source = "nonexistent_dir"
+    destination = test_dir_4
+    with raises(FileNotFoundError):
+        _doMove(source, destination)
+
+    # Non-existent destination directory
+    source = [test_file_2, test_dir_4]
+    destination = "nonexistent_dir"
+    with raises(FileNotFoundError):
+        _doMove(source, destination)
 
 
 @test("Basic remove operations.")
